Make stage/PR status columns editable, fix Add Item crash, support fractional quantities
- Production Stages and Purchase Requisitions tables now have clickable,
  role-gated status dropdowns (reusing StageCell; new PRStatusCell/PRStatusBadge)
- Item details: show an N/A symbol on each material row when the item is
  marked as needing no material
- Fix crash when adding an item via Data Management (missing purchaseReqs
  on the newly created item)
- OrderItem.outstandingQty is now Float to support fractional quantities

+++ src/app/api/items/[itemId]/route.ts
@@ -7,7 +7,7 @@ import { z } from "zod";
 const editSchema = z.object({
   description: z.string().optional(),
   productionOrderNo: z.string().optional(),
-  outstandingQty: z.number().int().min(0).optional(),
+  outstandingQty: z.number().min(0).optional(),
   drawingStatus: z.enum(["PENDING", "IN_PROGRESS", "DONE", "NA"]).optional(),
   carpentryStatus: z.enum(["PENDING", "IN_PROGRESS", "DONE", "NA"]).optional(),
   paintingStatus: z.enum(["PENDING", "IN_PROGRESS", "DONE", "NA"]).optional(),
